% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_seas.R
\name{regional_seas}
\alias{regional_seas}
\title{Create buffer divided by closest region}
\usage{
regional_seas(
  x,
  group,
  dist = units::set_units(30, km),
  density = units::set_units(0.1, 1/km)
)
}
\arguments{
\item{x}{Spatial polygon layer}

\item{group}{Character. The grouping variable for your subareas}

\item{dist}{distance from coasts of the buffer area. See ?sf::st_buffer}

\item{density}{density of points along the coastline.
(the higher, the more precise the region attribution)}
}
\description{
Create buffer divided by closest region
}
