% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunposition.R
\name{sun_position}
\alias{sun_position}
\title{Calculate the position of the sun according to
date and geographical position in wgs84
Found here:
http://stackoverflow.com/questions/8708048/position-of-the-sun-given-time-of-day-latitude-and-longitude}
\usage{
sun_position(year, month, day, hour = 12, min = 0, sec = 0,
  lat = 46.5, long = 6.5)
}
\arguments{
\item{year}{year}

\item{month}{month}

\item{day}{day}

\item{hour}{hour}

\item{min}{min}

\item{sec}{sec}

\item{lat}{lat}

\item{long}{long}
}
\description{
Calculate the position of the sun according to
date and geographical position in wgs84
Found here:
http://stackoverflow.com/questions/8708048/position-of-the-sun-given-time-of-day-latitude-and-longitude
}
